refactor: extract chat and sandbox panels
